# openapiDoc.JSON
{
  "openapi": "3.0.3",
  "info": {
    "title": "Groupomania social network",
    "description": "This is the documentation of Groupomania social network API.",
    "termsOfService": "http://groupomania.com/terms/",
    "contact": {
      "name": "API support",
      "url": "http://groupomania.com/contact/",
      "email": "contactAPI@groupomania.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
  "servers": [{
    "url": "http://localhost:{port}/{basePath}",
    "variables": {
      "port": {
        "default": "3001"
      },
      "basePath": {
        "default": "api"
      }
    }
  }],
...
  "paths": {
...
    "/reports": {
      "post": {
        "tags": ["reports"],
        "summary": "Create a report.",
        "description": "User creates a report.",
        "operationId": "createOneReport",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "User token.",
            "schema": {
              "$ref": "#/components/schemas/bearerToken"
            },
            "required": true
          }
        ],
        "requestBody": {
          "description": "Report properties. Exactly one of 'postId' or 'commentId' must be completed.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/reportPost"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Report created."
          },
          "400": {
            "description": "Bad request."
          },
          "401": {
            "description": "Unauthorized request."
          },
          "500": {
            "description": "Server error."
          }
        }
      },
      "get": {
        "tags": ["reports"],
        "summary": "Get all reports.",
        "description": "Admin gets all reports.",
        "operationId": "getAllReports",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "User token. Must be an admin one.",
            "schema": {
              "$ref": "#/components/schemas/bearerToken"
            },
            "required": true
          },
          {
            "name": "minDate",
            "in": "query",
            "description": "Get all reports who have been created after this date. Only the ISO 8601 format (YYYY-MM-DDTHH:mm:ss.sssZ) is supported.",
            "schema": {
              "type": "string",
              "example": "2022-13-01T14:48:02.809Z"
            }
          },
          {
            "name": "maxDate",
            "in": "query",
            "description": "Get all reports who have been created before this date. Only the ISO 8601 format (YYYY-MM-DDTHH:mm:ss.sssZ) is supported.",
            "schema": {
              "type": "string",
              "example": "2022-13-01T14:48:02.809Z"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Get a limited amount of reports.",
            "schema": {
              "type": "number",
              "example": 3
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Returns elements from oldest to newest ('asc') or from newest to oldest ('desc').",
            "schema": {
              "type": "string",
              "example": "asc"
            }
          },
          {
            "name": "fromUserId",
            "in": "query",
            "description": "Get all reports from a specific user.",
            "schema": {
              "type": "string",
              "example": "632dc722c75414645a459eb2"
            }
          },
          {
            "name": "fromPostId",
            "in": "query",
            "description": "Get all reports from a specific post.",
            "schema": {
              "type": "string",
              "example": "632efd854148fdc69cd7e652"
            }
          },
          {
            "name": "fromCommentId",
            "in": "query",
            "description": "Get all reports from a specific comment.",
            "schema": {
              "type": "string",
              "example": "632efd838148fdc69cd7e652"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ok.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/report"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request."
          },
          "403": {
            "description": "Forbidden request."
          },
          "500": {
            "description": "Server error."
          }
        }
      },
      "delete": {
        "tags": ["reports"],
        "summary": "Delete all reports.",
        "description": "Admin deletes all reports.",
        "operationId": "deleteAllReports",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "User token. Must be an admin one.",
            "schema": {
              "$ref": "#/components/schemas/bearerToken"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Ok."
          },
          "400": {
            "description": "Bad request."
          },
          "403": {
            "description": "Forbidden request."
          },
          "500": {
            "description": "Server error."
          }
        }
      }
    },
    "/reports/{reportId}": {
      "get": {
        "tags": ["reports"],
        "summary": "Get one report.",
        "description": "Admin gets properties of a report.",
        "operationId": "getOneReport",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "User token. Must be an admin one.",
            "schema": {
              "$ref": "#/components/schemas/bearerToken"
            },
            "required": true
          },
          {
            "name": "reportId",
            "in": "path",
            "description": "Id of report to return.",
            "schema": {
              "type": "string",
              "example": "632efdbb4198fdc69cd7e784"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Ok.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/report"
                }
              }
            }
          },
          "400": {
            "description": "Bad request."
          },
          "403": {
            "description": "Forbidden request."
          },
          "500": {
            "description": "Server error."
          }
        }
      },
      "delete": {
        "tags": ["reports"],
        "summary": "Delete one report.",
        "description": "Admin deletes a report.",
        "operationId": "deleteOneReport",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "description": "User token. Must be an admin one.",
            "schema": {
              "$ref": "#/components/schemas/bearerToken"
            },
            "required": true
          },
          {
            "name": "reportId",
            "in": "path",
            "description": "Id of report to delete.",
            "schema": {
              "type": "string",
              "example": "632efdbb4198fdc69cd7e784"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Ok."
          },
          "400": {
            "description": "Bad request."
          },
          "403": {
            "description": "Forbidden request."
          },
          "500": {
            "description": "Server error."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "report": {
        "title": "Report",
        "description": "Report schema. Exactly one of 'postId' or 'commentId' must be completed.",
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "example": "632efdbb4458fdc69cd7e784"
          },
          "postId": {
            "type": "string",
            "example": "632efd854148fdc69cd7e652"
          },
          "commentId": {
            "type": "string",
            "example": ""
          },
          "userId": {
            "type": "string",
            "example": "632dc722c75414645a459eb2"
          },
          "creationDate": {
            "type": "string",
            "example": "2022-13-01T14:48:02.809Z"
          },
          "__v": {
            "type": "number",
            "example": 0
          }
        },
        "required": [
          "_id",
          "userId",
          "creationDate",
          "__v"
        ]
      },
...
      "bearerToken": {
        "description": "Bearer token schema.",
        "type": "string",
        "example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiI2MzMxZDk1MWU3N2YxNDAxMmQxMjExNmEiLCJpc0FkbWluIjpmYWxzZSwiaWF0IjoxNjY0MjExMzEyLCJleHAiOjE2NjQyOTc3MTJ9.03h-xFcgHmwpYCHMZq36hnSAsHNS7TItboYmAliKSIk"
      },
...
      "reportPost": {
        "title": "ReportPost",
        "description": "Report schema used when user reports a post or comment. Exactly one of 'postId' or 'commentId' must be completed.",
        "type": "object",
        "properties": {
          "postId": {
            "type": "string",
            "example": "632efd854148fdc69cd7e892"
          },
          "commentId": {
            "type": "string",
            "example": "632efd854148fdc69cd7e892"
          }
        },
        "example": {
          "commentId" : "632efd854148fdc69cd7e652"
        }
      }
    }
  }
}